docker_geodjango/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
gunicorn==20.1.0
